+++ lib/exop/validation_checks.ex
@@ -16,11 +16,11 @@ defmodule Exop.ValidationChecks do
     * check_allow_nil/3
   """
 
+  alias Exop.TypeValidation
+
   @no_check_item :exop_no_check_item
 
   @type check_error :: %{(atom() | String.t()) => String.t()}
-
-  @known_types ~w(boolean integer float string tuple map struct list atom module function keyword uuid)a
 
   @doc """
   Returns an check_item's value from either a Keyword or a Map by an atom-key.
@@ -116,45 +116,11 @@ defmodule Exop.ValidationChecks do
     if check_item_present?(check_items, item_name) do
       check_item = get_check_item(check_items, item_name)
 
-      if Enum.member?(@known_types, check) do
-        do_check_type(check_item, check) || %{item_name => "has wrong type"}
-      else
-        true
-      end
+      TypeValidation.check_value(check_item, check) || %{item_name => "has wrong type"}
     else
       true
     end
   end
-
-  defp do_check_type(check_item, :boolean) when is_boolean(check_item), do: true
-  defp do_check_type(check_item, :integer) when is_integer(check_item), do: true
-  defp do_check_type(check_item, :float) when is_float(check_item), do: true
-  defp do_check_type(check_item, :string) when is_binary(check_item), do: true
-  defp do_check_type(check_item, :tuple) when is_tuple(check_item), do: true
-
-  defp do_check_type(%_{} = _check_item, :struct) do
-    IO.warn("type check with :struct is deprecated, please use :map instead")
-    true
-  end
-
-  defp do_check_type(_check_item, :struct) do
-    IO.warn("type check with :struct is deprecated, please use :map instead")
-    false
-  end
-
-  defp do_check_type(check_item, :map) when is_map(check_item), do: true
-  defp do_check_type(check_item, :list) when is_list(check_item), do: true
-  defp do_check_type(check_item, :atom) when is_atom(check_item), do: true
-  defp do_check_type(check_item, :function) when is_function(check_item), do: true
-  defp do_check_type([] = _check_item, :keyword), do: true
-  defp do_check_type([{atom, _} | _] = _check_item, :keyword) when is_atom(atom), do: true
-
-  defp do_check_type(check_item, :module) when is_atom(check_item) do
-    Code.ensure_loaded?(check_item)
-  end
-
-  defp do_check_type(check_item, :uuid) when is_binary(check_item), do: validate_uuid(check_item)
-  defp do_check_type(_, _), do: false
 
   @doc """
   Checks an item_name over numericality constraints.
@@ -249,7 +215,9 @@ defmodule Exop.ValidationChecks do
     check_number(number, item_name, {:less_than_or_equal_to, check_value})
   end
 
-  defp check_number(_number, _item_name, _), do: true
+  defp check_number(_number, item_name, {check, _check_value}) do
+    %{item_name => "unknwon check '#{check}'"}
+  end
 
   @doc """
   Checks whether an item_name is a memeber of a list.
@@ -351,46 +319,62 @@ defmodule Exop.ValidationChecks do
     actual_length = get_length(check_item)
 
     for {check, check_value} <- checks, into: [] do
-      case check do
-        :min -> check_min_length(item_name, actual_length, check_value)
-        :max -> check_max_length(item_name, actual_length, check_value)
-        :is -> check_is_length(item_name, actual_length, check_value)
-        :in -> check_in_length(item_name, actual_length, check_value)
-        _ -> true
-      end
+      check_length(check, item_name, actual_length, check_value)
     end
   end
 
   @spec get_length(any) :: pos_integer
-  defp get_length(param) when is_number(param), do: param
   defp get_length(param) when is_list(param), do: length(param)
   defp get_length(param) when is_binary(param), do: String.length(param)
-  defp get_length(param) when is_atom(param), do: param |> Atom.to_string() |> get_length
-  defp get_length(param) when is_map(param), do: param |> Map.to_list() |> get_length
+  defp get_length(param) when is_atom(param), do: param |> Atom.to_string() |> get_length()
+  defp get_length(param) when is_map(param), do: param |> Map.keys() |> get_length()
   defp get_length(param) when is_tuple(param), do: tuple_size(param)
-  defp get_length(_param), do: 0
+  defp get_length(_param), do: {:error, :wrong_type}
 
-  @spec check_min_length(atom() | String.t(), pos_integer, number) :: true | check_error
-  defp check_min_length(item_name, actual_length, check_value) do
+  @spec check_length(atom(), atom() | String.t(), pos_integer, number) :: true | check_error
+  defp check_length(_check, item_name, {:error, :wrong_type}, _check_value) do
+    %{item_name => "length check supports only lists, binaries, atoms, maps and tuples"}
+  end
+
+  defp check_length(:min, item_name, actual_length, check_value) do
+    check_length(:gte, item_name, actual_length, check_value)
+  end
+
+  defp check_length(:gte, item_name, actual_length, check_value) do
     actual_length >= check_value ||
       %{item_name => "length must be greater than or equal to #{check_value}"}
   end
 
-  @spec check_max_length(atom() | String.t(), pos_integer, number) :: true | check_error
-  defp check_max_length(item_name, actual_length, check_value) do
+  defp check_length(:gt, item_name, actual_length, check_value) do
+    actual_length > check_value ||
+      %{item_name => "length must be greater than #{check_value}"}
+  end
+
+  defp check_length(:max, item_name, actual_length, check_value) do
+    check_length(:lte, item_name, actual_length, check_value)
+  end
+
+  defp check_length(:lte, item_name, actual_length, check_value) do
     actual_length <= check_value ||
       %{item_name => "length must be less than or equal to #{check_value}"}
   end
 
-  @spec check_is_length(atom() | String.t(), pos_integer, number) :: true | check_error
-  defp check_is_length(item_name, actual_length, check_value) do
+  defp check_length(:lt, item_name, actual_length, check_value) do
+    actual_length < check_value ||
+      %{item_name => "length must be less than #{check_value}"}
+  end
+
+  defp check_length(:is, item_name, actual_length, check_value) do
     actual_length == check_value || %{item_name => "length must be equal to #{check_value}"}
   end
 
-  @spec check_in_length(atom() | String.t(), pos_integer, Range.t()) :: true | check_error
-  defp check_in_length(item_name, actual_length, check_value) do
+  defp check_length(:in, item_name, actual_length, check_value) do
     Enum.member?(check_value, actual_length) ||
       %{item_name => "length must be in range #{check_value}"}
+  end
+
+  defp check_length(check, item_name, _actual_length, _check_value) do
+    %{item_name => "unknown check '#{check}'"}
   end
 
   @doc """
@@ -504,44 +488,4 @@ defmodule Exop.ValidationChecks do
   defp validate_struct(%struct{}, struct, _item_name) when is_atom(struct), do: true
 
   defp validate_struct(_item, _check, item_name), do: %{item_name => "is not expected struct"}
-
-  @spec validate_uuid(binary()) :: boolean()
-  defp validate_uuid(
-         <<a1, a2, a3, a4, a5, a6, a7, a8, ?-, b1, b2, b3, b4, ?-, c1, c2, c3, c4, ?-, d1, d2, d3,
-           d4, ?-, e1, e2, e3, e4, e5, e6, e7, e8, e9, e10, e11, e12>>
-       ) do
-    <<c(a1), c(a2), c(a3), c(a4), c(a5), c(a6), c(a7), c(a8), ?-, c(b1), c(b2), c(b3), c(b4), ?-,
-      c(c1), c(c2), c(c3), c(c4), ?-, c(d1), c(d2), c(d3), c(d4), ?-, c(e1), c(e2), c(e3), c(e4),
-      c(e5), c(e6), c(e7), c(e8), c(e9), c(e10), c(e11), c(e12)>>
-  catch
-    :error -> false
-  else
-    _ -> true
-  end
-
-  defp validate_uuid(_), do: false
-
-  defp c(?0), do: ?0
-  defp c(?1), do: ?1
-  defp c(?2), do: ?2
-  defp c(?3), do: ?3
-  defp c(?4), do: ?4
-  defp c(?5), do: ?5
-  defp c(?6), do: ?6
-  defp c(?7), do: ?7
-  defp c(?8), do: ?8
-  defp c(?9), do: ?9
-  defp c(?A), do: ?a
-  defp c(?B), do: ?b
-  defp c(?C), do: ?c
-  defp c(?D), do: ?d
-  defp c(?E), do: ?e
-  defp c(?F), do: ?f
-  defp c(?a), do: ?a
-  defp c(?b), do: ?b
-  defp c(?c), do: ?c
-  defp c(?d), do: ?d
-  defp c(?e), do: ?e
-  defp c(?f), do: ?f
-  defp c(_), do: throw(:error)
 end