+++ test/operation_test.exs
@@ -17,22 +17,45 @@ defmodule OperationTest do
     %{name: :param2, opts: [type: :string]}
   ]
 
+  def operation_with_unknown_type do
+    defmodule WrongOperation do
+      use Exop.Operation
+
+      parameter :b, type: :unknown
+
+      def process(_), do: :ok
+    end
+  rescue
+    e ->
+      e
+  end
+
+  test "operation with unknown type check" do
+    assert %ArgumentError{
+             message:
+               "Unknown type check `:unknown` for parameter `:b` in module `OperationTest.WrongOperation`, " <>
+                 "supported type checks are `:boolean`, `:integer`, `:float`, `:string`, `:tuple`, `:struct`, " <>
+                 "`:map`, `:list`, `:atom`, `:function`, `:keyword`, `:module`, `:uuid`."
+           } = operation_with_unknown_type()
+  end
+
   test "defines contract/0" do
-    assert :functions |> Operation.__info__ |> Keyword.has_key?(:contract)
+    assert :functions |> Operation.__info__() |> Keyword.has_key?(:contract)
   end
 
   test "stores defined properties in a contract" do
-    assert Operation.contract |> is_list
-    assert Operation.contract |> List.first |> is_map
-    assert Enum.sort(Operation.contract) == Enum.sort(@valid_contract)
+    assert Operation.contract() |> is_list
+    assert Operation.contract() |> List.first() |> is_map
+    assert Enum.sort(Operation.contract()) == Enum.sort(@valid_contract)
   end
 
   test "defines run/1" do
-    assert :functions |> Operation.__info__ |> Keyword.has_key?(:run)
+    assert :functions |> Operation.__info__() |> Keyword.has_key?(:run)
   end
 
   test "process/1 takes a single param which is Map type" do
-    assert Operation.run(param1: 1, param2: "string") == {:ok, ["This is the process/1 params", %{param1: 1, param2: "string"}]}
+    assert Operation.run(param1: 1, param2: "string") ==
+             {:ok, ["This is the process/1 params", %{param1: 1, param2: "string"}]}
   end
 
   test "run/1: returns :validation_failed error when contract didn't pass validation" do
@@ -52,7 +75,7 @@ defmodule OperationTest do
       end
     end
 
-    assert DefOperation.run == {:ok, 999}
+    assert DefOperation.run() == {:ok, 999}
   end
 
   test "run/1: pass default value of required missed parameter (thus pass a validation)" do
@@ -84,37 +107,10 @@ defmodule OperationTest do
   end
 
   test "params/1: doesn't invoke a contract validation" do
-    assert Operation.process(param1: "not integer", param2: 777) == ["This is the process/1 params", [param1: "not integer", param2: 777]]
-  end
-
-  test "defined_params/0: returns params that were defined in the contract, filter out others" do
-    defmodule Def4Operation do
-      use Exop.Operation
-
-      parameter :a
-      parameter :b
-
-      def process(params) do
-        params |> defined_params
-      end
-    end
-
-    assert Def4Operation.run(a: 1, b: 2, c: 3) == {:ok, %{a: 1, b: 2}}
-  end
-
-  test "defined_params/0: respects defaults" do
-    defmodule Def5Operation do
-      use Exop.Operation
-
-      parameter :a
-      parameter :b, default: 2
-
-      def process(params) do
-        params |> defined_params
-      end
-    end
-
-    assert Def5Operation.run(a: 1, c: 3) == {:ok, %{a: 1, b: 2}}
+    assert Operation.process(param1: "not integer", param2: 777) == [
+             "This is the process/1 params",
+             [param1: "not integer", param2: 777]
+           ]
   end
 
   test "run/1: returns the last defined value for duplicated keys" do
@@ -143,7 +139,7 @@ defmodule OperationTest do
       end
     end
 
-    assert Def7Operation.run == {:interrupt, %{my_error: "oops"}}
+    assert Def7Operation.run() == {:interrupt, %{my_error: "oops"}}
   end
 
   test "interrupt/1: pass other exceptions" do
@@ -158,7 +154,7 @@ defmodule OperationTest do
       end
     end
 
-    assert_raise(RuntimeError, fn -> Def8Operation.run end)
+    assert_raise(RuntimeError, fn -> Def8Operation.run() end)
   end
 
   defmodule TruePolicy do
@@ -186,7 +182,7 @@ defmodule OperationTest do
       def process(_params), do: current_policy()
     end
 
-    assert Def9Operation.run == {:ok, {TruePolicy, :test}}
+    assert Def9Operation.run() == {:ok, {TruePolicy, :test}}
   end
 
   test "authorizes with provided policy" do
@@ -198,7 +194,7 @@ defmodule OperationTest do
       def process(_params), do: authorize(user: %TestUser{})
     end
 
-    assert Def10Operation.run == {:ok, :ok}
+    assert Def10Operation.run() == {:ok, :ok}
 
     defmodule Def11Operation do
       use Exop.Operation
@@ -208,7 +204,7 @@ defmodule OperationTest do
       def process(_params), do: authorize(user: %TestUser{})
     end
 
-    assert Def11Operation.run == {:error, {:auth, :test}}
+    assert Def11Operation.run() == {:error, {:auth, :test}}
   end
 
   test "operation invokation stops if auth failed" do
@@ -223,7 +219,7 @@ defmodule OperationTest do
       end
     end
 
-    assert Def12Operation.run == {:error, {:auth, :test}}
+    assert Def12Operation.run() == {:error, {:auth, :test}}
   end
 
   test "returns errors with malformed policy definition" do
@@ -243,8 +239,8 @@ defmodule OperationTest do
       def process(_params), do: authorize(%TestUser{})
     end
 
-    assert Def14Operation.run == {:error, {:auth, :unknown_policy}}
-    assert Def15Operation.run == {:error, {:auth, :unknown_policy}}
+    assert Def14Operation.run() == {:error, {:auth, :unknown_policy}}
+    assert Def15Operation.run() == {:error, {:auth, :unknown_policy}}
   end
 
   test "the last policy definition overrides previous definitions" do
@@ -257,49 +253,64 @@ defmodule OperationTest do
       def process(_params), do: current_policy()
     end
 
-    assert Def16Operation.run == {:ok, {FalsePolicy, :test}}
+    assert Def16Operation.run() == {:ok, {FalsePolicy, :test}}
   end
 
-  test "coerce option changes a parameter value (and after defaults resolving)" do
+  test "coerce function takes coerced param tuple and a map with all params" do
     defmodule Def17Operation do
       use Exop.Operation
 
-      parameter :a, default: 5, coerce_with: &__MODULE__.coerce/1
+      parameter :a, default: 5, coerce_with: &__MODULE__.coerce/2
       parameter :b
 
       def process(params), do: {params[:a], params[:b]}
 
-      def coerce(x), do: x * 2
+      def coerce({:a, a}, %{a: a, b: _}), do: a * 2
     end
 
     assert Def17Operation.run(b: 0) == {:ok, {10, 0}}
+  end
+
+  test "coerce option changes a parameter value (and after defaults resolving)" do
+    defmodule Def17aOperation do
+      use Exop.Operation
+
+      parameter :a, default: 5, coerce_with: &__MODULE__.coerce/2
+      parameter :b
+
+      def process(params), do: {params[:a], params[:b]}
+
+      def coerce({_, a}, _), do: a * 2
+    end
+
+    assert Def17aOperation.run(b: 0) == {:ok, {10, 0}}
   end
 
   test "coerce option changes a parameter value before validation" do
     defmodule Def18Operation do
       use Exop.Operation
 
-      parameter :a, numericality: %{greater_than: 0}, coerce_with: &__MODULE__.coerce/1
+      parameter :a, numericality: %{greater_than: 0}, coerce_with: &__MODULE__.coerce/2
 
       def process(params), do: params[:a]
 
-      def coerce(x), do: x * 2
+      def coerce({_, a}, _), do: a * 2
     end
 
     defmodule Def19Operation do
       use Exop.Operation
 
-      parameter :a, required: true, coerce_with: &__MODULE__.coerce/1
+      parameter :a, required: true, coerce_with: &__MODULE__.coerce/2
 
       def process(params), do: params[:a]
 
-      def coerce(_x), do: "str"
+      def coerce(_, _), do: "str"
     end
 
     defmodule Def20Operation do
       use Exop.Operation
 
-      parameter :a, func: &__MODULE__.validate/2, coerce_with: &__MODULE__.coerce/1
+      parameter :a, func: &__MODULE__.validate/2, coerce_with: &__MODULE__.coerce/2
       parameter :b, func: &__MODULE__.validate/1
 
       def process(params), do: params
@@ -309,6 +320,8 @@ defmodule OperationTest do
       def validate(x), do: x > 0
 
       def coerce(x), do: x + 1
+
+      def coerce({:a, a_value}, _received_params), do: a_value + 1
     end
 
     assert Def18Operation.run(a: 2) == {:ok, 4}
@@ -316,7 +329,9 @@ defmodule OperationTest do
 
     assert Def19Operation.run() == {:ok, "str"}
 
-    assert Def20Operation.run(a: -1, b: 0) == {:error, {:validation, %{a: ["isn't valid"], b: ["isn't valid"]}}}
+    assert Def20Operation.run(a: -1, b: 0) ==
+             {:error, {:validation, %{a: ["isn't valid"], b: ["isn't valid"]}}}
+
     assert Def20Operation.run(a: 0, b: 0) == {:error, {:validation, %{b: ["isn't valid"]}}}
     assert Def20Operation.run(a: 0, b: 1) == {:ok, %{a: 1, b: 1}}
   end
@@ -439,26 +454,34 @@ defmodule OperationTest do
     defmodule Def29Operation do
       use Exop.Operation
 
-      parameter :list_param, list_item: %{type: :string, length: %{min: 7}}, default: ["1234567", "7chars"]
+      parameter :list_param,
+        list_item: %{type: :string, length: %{min: 7}},
+        default: ["1234567", "7chars"]
 
       def process(params), do: {:ok, params[:list_param]}
     end
 
-    assert Def29Operation.run() == {:error, {:validation, %{"list_param[1]" => ["length must be greater than or equal to 7"]}}}
+    assert Def29Operation.run() ==
+             {:error,
+              {:validation, %{"list_param[1]" => ["length must be greater than or equal to 7"]}}}
   end
 
   test "list_item + coerce_with" do
     defmodule Def30Operation do
       use Exop.Operation
 
-      parameter :list_param, list_item: [type: :string, length: %{min: 7}], coerce_with: &__MODULE__.make_list/1
+      parameter :list_param,
+        list_item: [type: :string, length: %{min: 7}],
+        coerce_with: &__MODULE__.make_list/2
 
       def process(params), do: {:ok, params[:list_param]}
 
-      def make_list(_), do: ["1234567", "7chars"]
+      def make_list(_, _), do: ["1234567", "7chars"]
     end
 
-    assert Def30Operation.run() == {:error, {:validation, %{"list_param[1]" => ["length must be greater than or equal to 7"]}}}
+    assert Def30Operation.run() ==
+             {:error,
+              {:validation, %{"list_param[1]" => ["length must be greater than or equal to 7"]}}}
   end
 
   test "string-named parameters are allowed" do
@@ -471,9 +494,15 @@ defmodule OperationTest do
       def process(params), do: {:ok, params}
     end
 
-    assert Def31Operation.run() == {:error, {:validation, %{"a" => ["is required"], "b" => ["is required"]}}}
-    assert Def31Operation.run(%{"a" => 1, "b" => "2"}) == {:error, {:validation, %{"a" => ["has wrong type"], "b" => ["has wrong type"]}}}
-    assert Def31Operation.run(%{"a" => "1", b: 2}) == {:error, {:validation, %{"b" => ["is required"]}}}
+    assert Def31Operation.run() ==
+             {:error, {:validation, %{"a" => ["is required"], "b" => ["is required"]}}}
+
+    assert Def31Operation.run(%{"a" => 1, "b" => "2"}) ==
+             {:error, {:validation, %{"a" => ["has wrong type"], "b" => ["has wrong type"]}}}
+
+    assert Def31Operation.run(%{"a" => "1", b: 2}) ==
+             {:error, {:validation, %{"b" => ["is required"]}}}
+
     assert Def31Operation.run(%{"a" => "1", "b" => 2}) == {:ok, %{"a" => "1", "b" => 2}}
   end
 
@@ -487,8 +516,12 @@ defmodule OperationTest do
       def process(params), do: {:ok, params}
     end
 
-    assert Def32Operation.run() == {:error, {:validation, %{"a" => ["is required"], :b => ["is required"]}}}
-    assert Def32Operation.run(%{"a" => 1, b: "2"}) == {:error, {:validation, %{"a" => ["has wrong type"], :b => ["has wrong type"]}}}
+    assert Def32Operation.run() ==
+             {:error, {:validation, %{"a" => ["is required"], :b => ["is required"]}}}
+
+    assert Def32Operation.run(%{"a" => 1, b: "2"}) ==
+             {:error, {:validation, %{"a" => ["has wrong type"], :b => ["has wrong type"]}}}
+
     assert Def32Operation.run(%{"a" => "1"}) == {:error, {:validation, %{:b => ["is required"]}}}
     assert Def32Operation.run(%{"a" => "1", b: 2}) == {:ok, %{"a" => "1", :b => 2}}
   end
@@ -514,32 +547,16 @@ defmodule OperationTest do
     assert Def33Operation.run(a: 777) == {:ok, %{a: 777}}
   end
 
-  test "coerce_with can invoke a function with arity == 2 (passing param_name & param_value)" do
-    defmodule Def34Operation do
-      use Exop.Operation
-
-      parameter :a, type: :tuple, coerce_with: &__MODULE__.coerce_with_name/2
-
-      def process(params), do: params
-
-      def coerce_with_name(param_name, param_value) do
-        {List.duplicate(param_name, 2), param_value * 10}
-      end
-    end
-
-    assert Def34Operation.run(a: 1.1) == {:ok, %{a: {[:a, :a], 11}}}
-  end
-
   test "coerce_with respects an error-tuple result" do
     defmodule Def35Operation do
       use Exop.Operation
 
-      parameter :a, type: :integer, coerce_with: &__MODULE__.coerce/1
+      parameter :a, type: :integer, coerce_with: &__MODULE__.coerce/2
 
       def process(params), do: params
 
-      def coerce(1), do: {:error, :some_error}
-      def coerce(2), do: 2
+      def coerce({_, 1}, _), do: {:error, :some_error}
+      def coerce({_, 2}, _), do: 2
     end
 
     assert Def35Operation.run(a: 2) == {:ok, %{a: 2}}
@@ -560,14 +577,21 @@ defmodule OperationTest do
       assert Def36Operation.run(a: 1) == {:ok, %{a: 1}}
       assert Def36Operation.run(a: nil) == {:ok, %{a: nil}}
       assert Def36Operation.run(b: 1) == {:ok, %{b: 1}}
-      assert Def36Operation.run(b: nil) == {:error, {:validation, %{b: ["doesn't allow nil", "has wrong type"]}}}
+
+      assert Def36Operation.run(b: nil) ==
+               {:error, {:validation, %{b: ["doesn't allow nil", "has wrong type"]}}}
     end
 
     test "skips all checks" do
       defmodule Def37Operation do
         use Exop.Operation
 
-        parameter :a, type: :integer, numericality: [greater_than: 2], allow_nil: true, required: false
+        parameter :a,
+          type: :integer,
+          numericality: [greater_than: 2],
+          allow_nil: true,
+          required: false
+
         parameter :b, allow_nil: true, func: &__MODULE__.nil_check/2, required: false
 
         def nil_check(_, nil), do: {:error, :this_is_nil}
@@ -577,12 +601,15 @@ defmodule OperationTest do
 
       assert Def37Operation.run(a: nil) == {:ok, %{a: nil}}
       assert Def37Operation.run(a: 1) == {:error, {:validation, %{a: ["must be greater than 2"]}}}
-      assert Def37Operation.run(a: "1") == {:error, {:validation, %{a: ["not a number", "has wrong type"]}}}
+
+      assert Def37Operation.run(a: "1") ==
+               {:error, {:validation, %{a: ["not a number", "has wrong type"]}}}
+
       assert Def37Operation.run(b: nil) == {:ok, %{b: nil}}
     end
 
     test "required: false + allow_nil: false" do
-      defmodule Def45Operation do
+      defmodule Def37aOperation do
         use Exop.Operation
 
         parameter :a, required: false, allow_nil: false
@@ -590,9 +617,9 @@ defmodule OperationTest do
         def process(params), do: params
       end
 
-      assert Def45Operation.run(a: :a) == {:ok, %{a: :a}}
-      assert Def45Operation.run() == {:ok, %{}}
-      assert Def45Operation.run(a: nil) == {:error, {:validation, %{a: ["doesn't allow nil"]}}}
+      assert Def37aOperation.run(a: :a) == {:ok, %{a: :a}}
+      assert Def37aOperation.run() == {:ok, %{}}
+      assert Def37aOperation.run(a: nil) == {:error, {:validation, %{a: ["doesn't allow nil"]}}}
     end
   end
 
@@ -688,14 +715,120 @@ defmodule OperationTest do
     end
 
     assert Def44Operation.run(a: :a) == {:error, {:validation, %{a: ["has wrong type"]}}}
-    assert Def44Operation.run(a: []) == {:error, {:validation, %{"a[:b]" => ["is required"], "a[:c]" => ["is required"]}}}
-    assert Def44Operation.run(a: %{}) == {:error, {:validation, %{"a[:b]" => ["is required"], "a[:c]" => ["is required"]}}}
+
+    assert Def44Operation.run(a: []) ==
+             {:error, {:validation, %{"a[:b]" => ["is required"], "a[:c]" => ["is required"]}}}
+
+    assert Def44Operation.run(a: %{}) ==
+             {:error, {:validation, %{"a[:b]" => ["is required"], "a[:c]" => ["is required"]}}}
+
     assert Def44Operation.run(a: [b: :b, c: "c"]) == {:ok, %{a: [b: :b, c: "c"]}}
     assert Def44Operation.run(a: %{b: :b, c: "c"}) == {:ok, %{a: %{b: :b, c: "c"}}}
   end
 
+  describe "implicit inner" do
+    test "works with maps" do
+      defmodule Def45Operation do
+        use Exop.Operation
+
+        parameter :a, %{b: [type: :atom], c: [type: :string]}
+
+        def process(params), do: params
+      end
+
+      assert Def45Operation.run(a: :a) == {:error, {:validation, %{a: ["has wrong type"]}}}
+      assert Def45Operation.run(a: %{b: :b, c: "c"}) == {:ok, %{a: %{b: :b, c: "c"}}}
+
+      assert Def45Operation.run(a: %{}) ==
+               {:error, {:validation, %{"a[:b]" => ["is required"], "a[:c]" => ["is required"]}}}
+    end
+
+    test "works with keywords" do
+      defmodule Def46Operation do
+        use Exop.Operation
+
+        parameter :a, %{b: [type: :atom], c: [type: :string]}
+
+        def process(params), do: params
+      end
+
+      assert Def46Operation.run(a: :a) == {:error, {:validation, %{a: ["has wrong type"]}}}
+      assert Def46Operation.run(a: [b: :b, c: "c"]) == {:ok, %{a: [b: :b, c: "c"]}}
+
+      assert Def46Operation.run(a: []) ==
+               {:error, {:validation, %{"a[:b]" => ["is required"], "a[:c]" => ["is required"]}}}
+    end
+
+    test "doesnt work" do
+      defmodule Def47Operation do
+        use Exop.Operation
+
+        parameter :a, type: :atom
+
+        def process(params), do: params
+      end
+
+      assert Def47Operation.run(a: :a) == {:ok, %{a: :a}}
+      assert Def47Operation.run(a: "a") == {:error, {:validation, %{a: ["has wrong type"]}}}
+    end
+  end
+
+  defmodule(Def48Struct, do: defstruct([:a, :b]))
+
+  test "run/1 accepts a struct as params" do
+    defmodule Def48Operation do
+      use Exop.Operation
+
+      parameter :a, type: :integer
+      parameter :b, type: :string
+
+      def process(params), do: params
+    end
+
+    assert Def48Operation.run(a: 1, b: "1") == {:ok, %{a: 1, b: "1"}}
+    assert Def48Operation.run(%{a: 1, b: "1"}) == {:ok, %{a: 1, b: "1"}}
+    assert Def48Operation.run(%Def48Struct{a: 1, b: "1"}) == {:ok, %{a: 1, b: "1"}}
+
+    assert Def48Operation.run(%Def48Struct{a: "1", b: "1"}) ==
+             {:error, {:validation, %{a: ["has wrong type"]}}}
+  end
+
+  describe ":from option" do
+    test "makes an alias bw received param and desired param" do
+      defmodule Def49Operation do
+        use Exop.Operation
+
+        parameter :a, type: :integer, from: "a"
+        parameter :b, type: :string, from: :bB
+
+        def process(params), do: params
+      end
+
+      defmodule Def50Operation do
+        use Exop.Operation
+
+        parameter :a, type: :integer, from: :aA
+        parameter :b, type: :string, from: :bB
+
+        def process(params), do: params
+      end
+
+      assert Def49Operation.run(%{a: 1, b: "1"}) == {:ok, %{a: 1, b: "1"}}
+      assert Def49Operation.run(%{"a" => 1, b: "1"}) == {:ok, %{a: 1, b: "1"}}
+      assert Def49Operation.run(%{a: 1, bB: "1"}) == {:ok, %{a: 1, b: "1"}}
+      assert Def49Operation.run(%{"a" => 1, bB: "1"}) == {:ok, %{a: 1, b: "1"}}
+
+      assert Def49Operation.run(%{"a" => 1, bB: 1}) ==
+               {:error, {:validation, %{b: ["has wrong type"]}}}
+
+      assert Def50Operation.run(a: 1, b: "1") == {:ok, %{a: 1, b: "1"}}
+      assert Def50Operation.run(aA: 1, bB: "1") == {:ok, %{a: 1, b: "1"}}
+      assert Def50Operation.run(aA: 1, bB: 1) == {:error, {:validation, %{b: ["has wrong type"]}}}
+    end
+  end
+
   test ":inner check accepts opts as both map and keyword" do
-    defmodule Def46Operation do
+    defmodule Def51Operation do
       use Exop.Operation
 
       parameter :a, inner: [b: [type: :atom], c: [type: :string]]
@@ -703,7 +836,7 @@ defmodule OperationTest do
       def process(params), do: params
     end
 
-    defmodule Def47Operation do
+    defmodule Def52Operation do
       use Exop.Operation
 
       parameter :a, inner: %{b: [type: :atom], c: [type: :string]}
@@ -711,16 +844,41 @@ defmodule OperationTest do
       def process(params), do: params
     end
 
-    assert Def46Operation.run(a: :a) == {:error, {:validation, %{a: ["has wrong type"]}}}
-    assert Def46Operation.run(a: []) == {:error, {:validation, %{"a[:b]" => ["is required"], "a[:c]" => ["is required"]}}}
-    assert Def46Operation.run(a: %{}) == {:error, {:validation, %{"a[:b]" => ["is required"], "a[:c]" => ["is required"]}}}
-    assert Def46Operation.run(a: [b: :b, c: "c"]) == {:ok, %{a: [b: :b, c: "c"]}}
-    assert Def46Operation.run(a: %{b: :b, c: "c"}) == {:ok, %{a: %{b: :b, c: "c"}}}
+    assert Def51Operation.run(a: :a) == {:error, {:validation, %{a: ["has wrong type"]}}}
 
-    assert Def47Operation.run(a: :a) == {:error, {:validation, %{a: ["has wrong type"]}}}
-    assert Def47Operation.run(a: []) == {:error, {:validation, %{"a[:b]" => ["is required"], "a[:c]" => ["is required"]}}}
-    assert Def47Operation.run(a: %{}) == {:error, {:validation, %{"a[:b]" => ["is required"], "a[:c]" => ["is required"]}}}
-    assert Def47Operation.run(a: [b: :b, c: "c"]) == {:ok, %{a: [b: :b, c: "c"]}}
-    assert Def47Operation.run(a: %{b: :b, c: "c"}) == {:ok, %{a: %{b: :b, c: "c"}}}
+    assert Def51Operation.run(a: []) ==
+             {:error, {:validation, %{"a[:b]" => ["is required"], "a[:c]" => ["is required"]}}}
+
+    assert Def51Operation.run(a: %{}) ==
+             {:error, {:validation, %{"a[:b]" => ["is required"], "a[:c]" => ["is required"]}}}
+
+    assert Def51Operation.run(a: [b: :b, c: "c"]) == {:ok, %{a: [b: :b, c: "c"]}}
+    assert Def51Operation.run(a: %{b: :b, c: "c"}) == {:ok, %{a: %{b: :b, c: "c"}}}
+
+    assert Def52Operation.run(a: :a) == {:error, {:validation, %{a: ["has wrong type"]}}}
+
+    assert Def52Operation.run(a: []) ==
+             {:error, {:validation, %{"a[:b]" => ["is required"], "a[:c]" => ["is required"]}}}
+
+    assert Def52Operation.run(a: %{}) ==
+             {:error, {:validation, %{"a[:b]" => ["is required"], "a[:c]" => ["is required"]}}}
+
+    assert Def52Operation.run(a: [b: :b, c: "c"]) == {:ok, %{a: [b: :b, c: "c"]}}
+    assert Def52Operation.run(a: %{b: :b, c: "c"}) == {:ok, %{a: %{b: :b, c: "c"}}}
+  end
+
+  defmodule Def53Operation do
+    use Exop.Operation
+
+    parameter :a, type: :integer
+    parameter :b, type: :string
+
+    def process(params), do: params
+  end
+
+  test "process/1 takes only params defined in the contract" do
+    result = Def53Operation.run!(a: 1, b: "1", c: 2)
+
+    assert [:a, :b] = Map.keys(result)
   end
 end