+++ test/validation_test.exs
@@ -141,7 +141,7 @@ defmodule ValidationTest do
     {:error, {:validation, reasons}} = valid?(contract, received_params)
 
     assert %{
-      "list_param[0]" => ["has wrong type", "length must be greater than or equal to 7"],
+      "list_param[0]" => ["has wrong type", "length check supports only lists, binaries, atoms, maps and tuples"],
       "list_param[1]" => ["length must be greater than or equal to 7"],
       "list_param[2]" => ["has wrong type", "length must be greater than or equal to 7"]
     } == reasons